requirements.txt
Django==6.0.2
PyYAML==6.0.1
requests==2.32.3
beautifulsoup4==4.12.3
